late.c: Make work_memcpy_calibrate() more robust
Support the case where the initial 'count' leads to 'rmicro' being 0, by
retrying with an augmented count up to a limit that avoids overflow in
computations.  This however can only work on narrow circumstances, and
when that happens you'd better bump INT_MEMCPY_ITERATIONS.

Error out when the requested duration is too short to be representable
adequately by an (integral) number of iterations, instead of trying
indefinitely.

Error out when the requested duration is too long, without triggering
various overflows and having the calibration procedure loop infinitely.

Limit the number of iterations of the calibration loop to 8 by default.
Introduce '-a' to be able to change it from the command line.

Introduce '-l' (for leeway), the number of acceptable percents of error
in the calibration.  This is to account for fluctuating measurements for
a given number of iterations.  Set it to 5% by default.