late.c: Set absolute nice values, lift the "run as root" restriction
For calibration, we'll proceed with a run even if we can't elevate our
priority.  This is useful both for testing 'late' itself or on otherwise
idle machines for quick tests.

On FreeBSD only, we first try to raise to the largest (i.e., the
weakest) realtime priority.  On all systems, we then try to set the nice
value to -20.  If all fails, we print a warning instead of exiting.

'-n' now takes an absolute nice value.  When it is specified, and
setting the nice value fails, we exit loudly instead of silently continuing.