late.c: Calibration: Be more precise, avoid overflows
Increase SCALE to be greater than the precision limit enforced by the
"leeway" ('-l'), else the calibration process may not converge as the
correction ratio applied to the iteration count is not precise enough,
making it remain the same or jump back and forth around the proper
value.

In order to achieve that, given the order of magnitude of the number of
microseconds passed in, widen the type of all microseconds counts to
'uint64_t' (else overflows may occur with not too far-fetched values).

Fix the comparison between the expected duration and the real one. SCALE
was not really effective there as we were dividing by it before the
comparison (and were using SCALE + 1 to compensate for C's integer
division's truncation, which has been removed also).

Fix overflow handling when computing 'new_count'.  There is no need to
check for room to multiply 'count' by SCALE in some cases, as this is
not what 'count' is going to be multiplied with.  Just check if the
computation of 'new_count' actually overflows.  Add a sanity check about
SCALE being big enough (in case SCALE or the '-l' scale are later
changed in contradiction to the rule mentioned in the first paragraph
above).

While here, print the ratio to be applied to the current 'count' in
a more intelligible way.