late.c: Turn 'is_count' into an unsigned int
This is a count, so should never get into negative territory.  In most
circumstances, it will not overflow, but could be if left running for
a long time with very small values passed as sleep time and work
iterations.  In that unlikely case, just stop updating the count and
total time so that the reported values of count and average stay
meaningful (although they won't take into account recent history).