late.c: Make specifying '-s' along with '-w' mandatory
There is no real rationale for having a sleep time default. Most tests
assume a cell of 100ms, making the sleep time dependent on the chosen
work time.

All existing tests are already compatible with this change.

While here, remove the useless test on 'wflag' in the execution pattern
loop, as it cannot be 'false' at this point.

While here, remove the superfluous 'else' when determining the mode and
checking presence of the related options.