late.c: gettimeofday(), 'struct timeval' => clock_gettime(), 'struct timespec'
This change is mostly in preparation for being able to call
clock_gettime() from the finished() signal handler, as it is required by
POSIX to be async-signal-safe.  gettimeofday() has been obsoleted long
ago and does not offer such a guarantee.

While here, encapsulate calls to gettimeofday()/clock_gettime() into
convenient functions, and have a separate function to get the clock
without overhead at the expense of a loss of precision, which is
a better alternative for supporting option '-r'.

The different CLOCK_* constants should be detected and correctly acted
upon (in a future-proof manner, provided OSes do not begin to do weird
things).  Detection of the presence of clock_nanosleep() is not
automatic; platforms supporting it must define HAS_CLOCK_NANOSLEEP.
This has been done for FreeBSD and Linux.