late.c: Calibration: Add a stabilization phase
During this phase, we repeat the calibration work to average the result.
If, during this repetition, one calibration results in an execution time
out of the leeway, wexit the stabilization phase and continue with
regular feedback loops.

Introduce a new parameter, which is the number of repetitions (and thus,
samples to compute the average) for the stabilization phase, and set it
to 3 by default.

While here, choose a clearer error message on work iterations overflow.